Refactor notifications to single hook
