+++ client/my-sites/domains/domain-management/list/bulk-site-domains.tsx
@@ -2,7 +2,7 @@ import page from '@automattic/calypso-router';
 import { useSiteDomainsQuery } from '@automattic/data-stores';
 import { DomainsTable, ResponseDomain } from '@automattic/domains-table';
 import { useTranslate } from 'i18n-calypso';
-import { useEffect, useMemo, useState } from 'react';
+import { useMemo, useState } from 'react';
 import SiteAddressChanger from 'calypso/blocks/site-address-changer';
 import DocumentHead from 'calypso/components/data/document-head';
 import InlineSupportLink from 'calypso/components/inline-support-link';
@@ -17,7 +17,6 @@ import {
 	showUpdatePrimaryDomainErrorNotice,
 	showUpdatePrimaryDomainSuccessNotice,
 } from 'calypso/state/domains/management/actions';
-import { successNotice } from 'calypso/state/notices/actions';
 import { setPrimaryDomain } from 'calypso/state/sites/domains/actions';
 import { hasDomainCredit as hasDomainCreditSelector } from 'calypso/state/sites/plans/selectors';
 import { isSupportSession } from 'calypso/state/support/selectors';
@@ -37,6 +36,7 @@ import EmptyDomainsListCard from './empty-domains-list-card';
 import GoogleDomainOwnerBanner from './google-domain-owner-banner';
 import ManageAllDomainsButton from './manage-all-domains-button';
 import OptionsDomainButton from './options-domain-button';
+import { useNewDomainsNotification } from './use-new-domains-notification';
 import { usePurchaseActions } from './use-purchase-actions';
 import { filterOutWpcomDomains } from './utils';
 import './style.scss';
@@ -118,24 +118,7 @@ export default function BulkSiteDomains( props: BulkSiteDomainsProps ) {
 		}
 	};
 
-	// Domain purhcases provide a new domain flag on checkout completion
-	// e.g. /domains/manage/example.wordpress.com?new-domains=2
-	const queryParams = new URLSearchParams( window.location.search );
-	const newDomains = queryParams.get( 'new-domains' );
-	useEffect( () => {
-		if ( newDomains ) {
-			dispatch(
-				successNotice(
-					translate( 'Your domain is being setup.', 'Your domains are being set up.', {
-						count: parseInt( newDomains ),
-					} ),
-					{
-						duration: 10000,
-					}
-				)
-			);
-		}
-	}, [ newDomains, dispatch, translate ] );
+	useNewDomainsNotification();
 
 	return (
 		<>
